Try out a trait for checking whether two instances are same type
Motivated by seeing a `same_type` function implemented for `toml::Value`
in the `toml` crate. (And `toml::Value` is a perfect example of where
you would want to use a type federation.)

Weakness as currently written
-----------------------------
Highly inelegant.

Should possibly be able to be used to compare instance of different
federations (as long as the types in the first federation form a subset
of the types in the second federation...not yet sure of exactly which
constraints would be useful...perhaps a superset would also be useful?).
For example:

```rust
  let a: Fed2<String, u8> = 123u8.into();
  let b: Fed3<String, u8, bool> = String::from("abc").into();
  let c: Fed3<String, u8, bool> = 0u8.into();

  assert!(!a.is_same_type(&b));
  assert!(a.is_same_type(&c));
```