Include derive capability for Fed0 & Fed1 & update docs
Had neglected to include the auto-derive directive introduced in 0.4 for
Fed0 and Fed1. That is now fixed.

I also updated the README to include an example of using the auto-derive
capability in `init_fed!`.

The docs property of Cargo.toml now points to the non-version specified
URL of docs.rs.